Add DirectAdmin option
I have written a documentation entry where it states how to host Jexactyl in DirectAdmin without difficulty. This usually is for people who have websites in one server and don't want to change the HTTP(s) ports for each webserver running, one running DA and the other one the panel.